Move map component to a separated file
