+++ web-vite/src/main.tsx
@@ -1,4 +1,5 @@
 import { render } from "preact";
 import { App } from "./app.tsx";
+import "./index.css";
 
 render(<App />, document.getElementById("app")!);
